Rename folder : list-battles -> list-battle-sessions
