+++ src/frontend/list-battle-sessions/actions/delete-battle-session.ts
