+++ src/frontend/list-battle-sessions/actions/load-battle-sessions.ts
