+++ src/frontend/list-battle-sessions/actions/open-deletion-modal.ts
