+++ src/frontend/list-battle-sessions/components/battle-sessions-list.tsx
