+++ src/frontend/list-battle-sessions/list-battles-container.ts
