+++ src/frontend/list-battle-sessions/reducers/reducers.ts
