+++ src/frontend/list-battle-sessions/types/modal.ts
